+++ src/br/com/constants/HasPhone.java
@@ -2,8 +2,8 @@ package br.com.constants;
 
 public interface HasPhone {
 
-    public Integer getPhone();
+    public Long getPhone();
 
-    public void setPhone(Integer phone);
+    public void setPhone(Long phone);
 
 }
